CI_CD/Apps/traffic_stats/requirements.txt
pymongo
kafka-python
python-dotenv
termcolor
pyshark
confluent_kafka